% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividi_loot.R
\name{dividi_loot}
\alias{dividi_loot}
\title{Effettua le divisioni di monete di varia tipologia}
\usage{
dividi_loot(MO = 0, MA = 0, MR = 0, MP = 0, ME = 0, n = 7)
}
\arguments{
\item{MO}{monete oro}

\item{MA}{monete argento}

\item{MR}{monete rame}

\item{MP}{monete platino}

\item{ME}{monete elettro}

\item{n}{numero di PG per i quale fare la ripartizione}
}
\description{
Effettua le divisioni di monete di varia tipologia
}
