% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{monsters}
\alias{monsters}
\title{Monsters list}
\format{An object of class \code{list} (inherits from \code{monsterList}) of length 325.}
\usage{
monsters
}
\description{
Monsters list
}
\keyword{datasets}
