% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spellbook_backup.R
\name{spellbook_backup}
\alias{spellbook_backup}
\title{Calcola il costo di un backup di spellbook in base alle magie trascritte}
\usage{
spellbook_backup(copied_spells = NULL, add_blank_spellbook = TRUE)
}
\arguments{
\item{copied_spells}{vector with the number of trascribed}
}
\description{
Calcola il costo di un backup di spellbook in base alle magie trascritte
}
\examples{
spellbook_backup(c(6,4,2)) # 6 lev1 spells, 4 lev2 spells, 2 lev3 spells

}
