% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{spells}
\alias{spells}
\title{Spells list}
\format{An object of class \code{list} (inherits from \code{spellList}) of length 460.}
\usage{
spells
}
\description{
Spells list
}
\keyword{datasets}
