% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_cv.R
\name{month_to_cv}
\alias{month_to_cv}
\title{Funzione che trasforma un mese numerico nel nome in CV}
\usage{
month_to_cv(x)
}
\arguments{
\item{x}{number of month}
}
\description{
Funzione che trasforma un mese numerico nel nome in CV
}
