% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spellbook_backup.R
\name{spellbook_backup}
\alias{spellbook_backup}
\title{Calcola il costo di un backup di spellbook in base alle magie trascritte}
\usage{
spellbook_backup(
  n_full_cost = NULL,
  n_half_cost = NULL,
  add_blank_spellbook = TRUE
)
}
\arguments{
\item{n_full_cost}{vector with the number of trascribed spells
which are payed at full cost (eg if evocator, all but evocation
spells)}

\item{n_half_cost}{vector with the number of trascribed spells
which are payed at half cost (eg if evocator, only evocation
spells)}
}
\description{
Calcola il costo di un backup di spellbook in base alle magie trascritte
}
\examples{
# Aelar 5th level
spellbook_backup(c(9,4,2), c(3,0,2))

}
